LPS-57254 Comments - Rename to "comment"
