+++ modules/apps/comments/comments-test/bnd.bnd
@@ -1,5 +1,5 @@
-Bundle-Name: Liferay Comments Test
-Bundle-SymbolicName: com.liferay.comments.test
+Bundle-Name: Liferay Comment Comments Test
+Bundle-SymbolicName: com.liferay.comment.comments.test
 Bundle-Version: 1.0.0
 Include-Resource:\
 	classes,\
